Minor: Move MemoryCatalog*Provider into a module, improve comments
